moved IndexedDB code into own files
